image_generation/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
datasets==4.0.0
diffusers==0.35.1
dill==0.3.8
eval_type_backport==0.2.2
fairscale==0.4.13
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.8
hpsv2==1.2.0
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
image-reward==1.5
importlib_metadata==8.7.0
inflect==6.0.4
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-clip==1.0.1
packaging==25.0
pandas==2.3.2
peft==0.17.0
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
pyarrow==21.0.0
pydantic==1.10.13
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
sentry-sdk==2.35.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
termcolor==3.1.0
timm==0.6.13
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.3
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
